更新步兵和英雄代码，英雄代码未烧录测试

+++ components/protocol/ui.hpp
@@ -16,7 +16,7 @@ public:
 
     uint8_t robot_id = 0;
 
-    uint8_t set_bullet_frequency = 10; // 弹丸发射频率
+    int16_t set_bullet_frequency = 10; // 弹丸发射频率
     uint8_t set_gimbal_is_imu_mode = 1; // 云台模式，是imu模式为1，否则为0
     uint8_t set_chassis_max_speed = 1; // 底盘平移速度
     uint8_t set_is_locked = 0; // 自瞄是否锁定
@@ -29,7 +29,7 @@ public:
     uint8_t set_chassis_power_limit = 0;
 
 private:
-    uint8_t set_bullet_frequency_last = 10; // 弹丸发射频率
+    int16_t set_bullet_frequency_last = 10; // 弹丸发射频率
     uint8_t set_gimbal_is_imu_mode_last = 1; // 云台模式，是imu模式为1，否则为0
     uint8_t set_chassis_max_speed_last = 1; // 底盘平移速度
     uint8_t set_is_locked_last = 0; // 自瞄是否锁定
